Move failure tests to a new module.
